+++ packages/cli/src/index.ts
@@ -12,7 +12,9 @@ import {
   initProject,
   exportSite,
   serveSite,
+  generateCodeForProject,
   BuildCLIOptions,
+  GenerateCodeCLIOptions,
 } from "@featurevisor/core";
 
 process.on("unhandledRejection", (reason, p) => {
@@ -134,6 +136,29 @@ async function main() {
     .example("$0 site export", "generate static site with project data")
     .example("$0 site serve", "serve already exported site locally")
     .example("$0 site serve -p 3000", "serve in a specific port")
+
+    /**
+     * Generate code
+     */
+    .command({
+      command: "generate-code",
+      handler: function (options) {
+        const projectConfig = requireAndGetProjectConfig(rootDirectoryPath);
+
+        try {
+          generateCodeForProject(
+            rootDirectoryPath,
+            projectConfig,
+            options as unknown as GenerateCodeCLIOptions,
+          );
+        } catch (e) {
+          console.error(e.message);
+          process.exit(1);
+        }
+      },
+    })
+    .example("$0 generate-code", "generate code from YAMLs")
+    .example("$0 generate-code --language typescript --out-dir ./src", "")
 
     /**
      * Options