+++ packages/vue/jest.config.js
@@ -1,0 +1,7 @@
+module.exports = {
+  preset: "ts-jest",
+  bail: true,
+
+  testEnvironment: "jsdom",
+  setupFilesAfterEnv: ["@testing-library/jest-dom/extend-expect", "./jest.setup.js"],
+};